Enhance note to self support, add "send" receiver
Creating a new Instance from a name will now reuse an existing Task if
there's already one out there, and an existing Instance if there's one
that's incomplete.

Also added support for using it as a "Send" or "Share" receiver.
Creating new things to do is now easier than ever.